+++ test/daemon_test.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
@@ -211,148 +210,8 @@ main() {
     Directory.current = 'example';
 
     await runTestsOnAll(
-        daemonClient, devices, emulators, configInfo, imageProcessor);
+        daemonClient, devices, emulators, config, screens, imageProcessor);
     // allow other tests to continue
     Directory.current = origDir;
   }, timeout: Timeout(Duration(minutes: 4)));
-}
-
-Future runTestsOnAll(DaemonClient daemonClient, List devices, List emulators,
-    Map configInfo, ImageProcessor imageProcessor) async {
-  List deviceNames = getAllDevices(configInfo);
-  final locales = configInfo['locales'];
-  final stagingDir = configInfo['staging'];
-  final testPaths = configInfo['tests'];
-
-  for (final deviceName in deviceNames) {
-    DeviceType deviceType;
-    // look for matching device first
-    final device = devices.firstWhere((device) {
-      if (device['platform'] == 'ios') {
-        deviceType = DeviceType.ios;
-        if (device['emulator']) {
-          // running ios simulator
-          return device['name'] == deviceName;
-        } else {
-          // real ios device
-          return device['model'].contains(deviceName);
-        }
-      } else {
-        deviceType = DeviceType.android;
-        if (device['emulator']) {
-          // running android emulator
-          return findDeviceEmulator(emulators, device['id'])['name'] ==
-              deviceName;
-        } else {
-          // real android device
-          return device['name'] == deviceName;
-        }
-      }
-    }, orElse: () => null);
-
-    String deviceId;
-    Map emulator = null;
-    Map simulator = null;
-    if (device != null) {
-      deviceId = device['id'];
-    } else {
-      // if no matching device, look for matching android emulator
-      emulator = findEmulator(emulators, deviceName);
-      if (emulator != null) {
-        deviceType = DeviceType.android;
-        final emulatorId = emulator['id'];
-        print('Starting $deviceName...');
-//        daemonClient.verbose = true;
-        await daemonClient.launchEmulator(emulatorId);
-//        daemonClient.verbose = false;
-        deviceId = findAndroidDeviceId(emulatorId);
-        print('... $deviceName started.');
-      } else {
-        // if no matching android emulator, look for matching ios simulator
-        deviceType = DeviceType.ios;
-        simulator = getHighestIosSimulator(getIosSimulators(), deviceName);
-        deviceId = simulator['udid'];
-        print('Starting $deviceName...');
-//        daemonClient.verbose = true;
-        cmd('xcrun', ['simctl', 'boot', deviceId]);
-//        daemonClient.verbose = false;
-        print('... $deviceName started.');
-      }
-    }
-
-    for (final locale in locales) {
-      // set locale if android device or emulator
-      if ((device != null && device['platform'] != 'ios') ||
-          (device == null && emulator != null)) {
-        // a running android device or emulator
-        final deviceLocale = androidDeviceLocale(deviceId);
-        print('android device or emulator locale=$deviceLocale');
-        if (locale != deviceLocale) {
-          print('Changing locale from $deviceLocale to $locale...');
-          daemonClient.verbose = true;
-          setAndroidLocale(deviceId, deviceName, locale);
-          daemonClient.verbose = false;
-          print('... locale change complete.');
-        }
-      }
-
-      // set locale if ios simulator
-      if ((device != null &&
-              device['platform'] == 'ios' &&
-              device['emulator']) ||
-          (device == null && simulator != null)) {
-        // a running simulator
-        final deviceLocale = iosSimulatorLocale(deviceId);
-        print('simulator locale=$deviceLocale');
-        if (locale != deviceLocale) {
-          print('Changing locale from $deviceLocale to $locale');
-          daemonClient.verbose = true;
-          cmd('xcrun', ['simctl', 'shutdown', deviceId]);
-          await streamCmd('$stagingDir/resources/script/simulator-controller',
-              [deviceId, 'locale', locale]);
-          cmd('xcrun', ['simctl', 'boot', deviceId]);
-          daemonClient.verbose = true;
-          print('...locale change complete.');
-        }
-      }
-
-      // issue warning if ios device
-      if ((device != null &&
-          device['platform'] == 'ios' &&
-          !device['emulator'])) {
-        // a running ios device
-        print('Warning: the locale of an ios device cannot be changed.');
-      }
-
-      // run tests
-      for (final testPath in testPaths) {
-        print('Running $testPath on \'$deviceName\' in locale $locale...');
-        await streamCmd('flutter', ['-d', deviceId, 'drive', testPath]);
-
-        // process screenshots
-        await imageProcessor.process(deviceType, deviceName, locale);
-      }
-    }
-
-    // if an emulator was started, shut it down
-    if (emulator != null) {
-      await shutdownAndroidEmulator(deviceId, emulator['name']);
-    }
-    if (simulator != null) {
-      print('Waiting for \'$deviceName\' to shutdown...');
-      cmd('xcrun', ['simctl', 'shutdown', deviceId]);
-      print('... \'$deviceName\' shutdown complete.');
-    }
-  }
-}
-
-Map findEmulator(List emulators, String emulatorName) {
-  return emulators.firstWhere((emulator) => emulator['name'] == emulatorName,
-      orElse: () => null);
-}
-
-Map findDeviceEmulator(List emulators, String deviceId) {
-  return emulators.firstWhere(
-      (emulator) => emulator['id'] == getAndroidEmulatorId(deviceId),
-      orElse: () => null);
 }
